refactor: extract worldStateManager into modular structure (#906)
