+++ src/lib/world/events/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @fileoverview World Events Module
+ *
+ * Story checkpoint and major event management.
+ */
+export * from './majorEvents';
+export * from './storyCheckpoints';
