+++ src/lib/world/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @fileoverview World State Management Module
+ *
+ * Complete world state management system including relationships,
+ * events, checkpoints, and state synchronization.
+ */
+// Re-export everything from worldStateManager (main public API)
+export * from './worldStateManager';
+
+// Re-export sub-modules for direct access if needed
+export * as events from './events';
+export * as relationships from './relationships';
+export * as utils from './utils';
