+++ src/lib/world/relationships/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @fileoverview Relationship Management Module
+ *
+ * NPC and character relationship tracking and updates.
+ */
+export * from './npcRelationships';
+export * from './characterRelationships';
