+++ src/lib/world/utils/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @fileoverview World State Utilities
+ *
+ * Shared utilities for world state management including deduplication,
+ * normalization, and merging functions.
+ */
+export * from './deduplication';
+export * from './normalization';
+export * from './merging';
